Refactor project for multi-module maven build
A full refactoring of the project to take advantage of Maven.

Refactoring includes naming, multi-module build, and more. Updated the
URL for the BackToTheFuture example to use an archived version of the
RottenTomatoes website at archive.org at the time of the original commit
. Changes at RottenTomatoes prevented this from working.

Maven GAV was refactored to use case normalization.

The build creates an executable JAR (with dependencies).

Update the SNAPSHOT version.